refactor(web): merge-dispute-features-to-dispute-kit-regsitry
