+++ web/src/dispute-kits/disputeFeature.ts
@@ -1,39 +1,10 @@
-import { DisputeKits } from "src/dispute-kits";
+import { DisputeKitConfig, DisputeKits } from "src/dispute-kits";
 
 import { isProductionDeployment } from "../consts/index";
 
 import { DisputeKitDataMap, GatedDisputeKitData } from "./prepareArbitratorExtradata";
-
-export enum Group {
-  Voting = "Voting",
-  Eligibility = "Eligibility",
-}
-
-/** A single feature, grouped into categories. has to be atomic.
- * For gated, we split them into atomic erc20 and erc1155 */
-export enum Features {
-  ShieldedVote = "shieldedVote",
-  ClassicVote = "classicVote",
-  UniversityVote = "universityVote",
-  ClassicEligibility = "classicEligibility",
-  GatedErc20 = "gatedErc20",
-  GatedErc1155 = "gatedErc1155",
-  ArgentinaConsumerProtection = "argentinaConsumerProtection",
-}
-
-/** Group of features (like radio buttons per category) */
-export type FeatureGroups = Record<string, Features[]>;
-
-/** Definition of a dispute kit */
-export interface DisputeKit {
-  id: DisputeKits;
-  /**
-   * The feature sets this kit supports.
-   * Each array represents a valid configuration, and has to be 1:1,
-   * if either subset matches the selected feature array this dispute kit is selected
-   */
-  featureSets: Features[][];
-}
+import { DISPUTE_KIT_REGISTRY } from "./registry";
+import { FeatureGroups, Features, Group } from "./types";
 
 // groups
 // withing a group only one feature can be selected, we deselect the other one when a new one is selected
@@ -50,49 +21,15 @@ export const featureGroups: FeatureGroups = {
   ],
 };
 
-// dispute kits
-// each array is a unique match, for multiple combinations, add more arrays.
-// TODO: move to dk cofig registry
-// we need this to structure disputekit data for the dispute, not sure how to move this to config yet
-export const disputeKits: DisputeKit[] = [
-  {
-    id: DisputeKits.Classic,
-    featureSets: [[Features.ClassicVote, Features.ClassicEligibility]],
-  }, // strict
-  { id: DisputeKits.Shutter, featureSets: [[Features.ShieldedVote, Features.ClassicEligibility]] }, // strict
-  {
-    id: DisputeKits.Gated,
-    // strictly keep the common feature in front and in order.
-    featureSets: [
-      [Features.ClassicVote, Features.GatedErc20],
-      [Features.ClassicVote, Features.GatedErc1155],
-    ],
-  },
-  {
-    id: DisputeKits.GatedShutter,
-    featureSets: [
-      [Features.ShieldedVote, Features.GatedErc20],
-      [Features.ShieldedVote, Features.GatedErc1155],
-    ],
-  },
-  {
-    id: DisputeKits.ArgentinaConsumerProtection,
-    featureSets: [[Features.ClassicVote, Features.ArgentinaConsumerProtection]],
-  },
-  {
-    id: DisputeKits.ClassicUniversity,
-    featureSets: [[Features.UniversityVote, Features.ClassicEligibility]],
-  },
-];
-
 // excluded on mainnet, we can later update it to an array if we want to have more deployment specific kits
 const UNIVERSITY_DISPUTE_KIT_ID = DisputeKits.ClassicUniversity;
 
+const disputeKits = Object.values(DISPUTE_KIT_REGISTRY);
 /**
  * Dispute kits available for the current deployment.
  * University DK is excluded on mainnet.
  */
-export const getDisputeKitsForDeployment = (): DisputeKit[] =>
+export const getDisputeKitsForDeployment = (): DisputeKitConfig[] =>
   isProductionDeployment() ? disputeKits.filter((kit) => kit.id !== UNIVERSITY_DISPUTE_KIT_ID) : disputeKits;
 
 /**
@@ -107,7 +44,7 @@ export const getFeatureGroupsForDeployment = (): FeatureGroups =>
       }
     : featureGroups;
 
-export function resolveInitialFeatureSet(kit: DisputeKit, disputeKitData?: DisputeKitDataMap[DisputeKits]) {
+export function resolveInitialFeatureSet(kit: DisputeKitConfig, disputeKitData?: DisputeKitDataMap[DisputeKits]) {
   switch (kit.id) {
     case DisputeKits.Classic:
     case DisputeKits.Shutter:
@@ -166,7 +103,7 @@ export function toggleFeature(selected: Features[], feature: Features, groups: F
 /**
  * Find dispute kits that match the given selection
  */
-export function findMatchingKits(selected: Features[], kits: DisputeKit[]): DisputeKit[] {
+export function findMatchingKits(selected: Features[], kits: DisputeKitConfig[]): DisputeKitConfig[] {
   return kits.filter((kit) =>
     kit.featureSets.some(
       (set) => arraysEqual(set, selected) // strict exact match
@@ -188,7 +125,7 @@ export function findMatchingKits(selected: Features[], kits: DisputeKit[]): Disp
  *
  * @returns  A corrected selection that is guaranteed to be valid.
  */
-export function ensureValidSmart(selected: Features[], groups: FeatureGroups, kits: DisputeKit[]): Features[] {
+export function ensureValidSmart(selected: Features[], groups: FeatureGroups, kits: DisputeKitConfig[]): Features[] {
   // --- Helper: checks if a candidate is valid or could still become valid ---
   function isValidOrPrefix(candidate: Features[]): boolean {
     return (
@@ -219,7 +156,7 @@ export function ensureValidSmart(selected: Features[], groups: FeatureGroups, ki
 }
 
 // Checks if the candidate if selected, can still lead to a match
-function canStillLeadToMatch(candidate: Features[], kits: DisputeKit[]): boolean {
+function canStillLeadToMatch(candidate: Features[], kits: DisputeKitConfig[]): boolean {
   return kits.some((kit) =>
     kit.featureSets.some((set) =>
       // candidate must be subset of this set
@@ -232,7 +169,11 @@ function canStillLeadToMatch(candidate: Features[], kits: DisputeKit[]): boolean
  * Compute which features should be disabled,
  * given the current selection.
  */
-export function getDisabledOptions(selected: Features[], groups: FeatureGroups, kits: DisputeKit[]): Set<Features> {
+export function getDisabledOptions(
+  selected: Features[],
+  groups: FeatureGroups,
+  kits: DisputeKitConfig[]
+): Set<Features> {
   const disabled = new Set<Features>();
 
   // If nothing is selected => allow all
@@ -262,7 +203,7 @@ export function getDisabledOptions(selected: Features[], groups: FeatureGroups, 
  */
 export function getVisibleFeaturesForCourt(
   supportedKits: number[],
-  allKits: DisputeKit[],
+  allKits: DisputeKitConfig[],
   groups: FeatureGroups
 ): FeatureGroups {
   // Get supported kits for this court
